# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Savezy API
  version: 1.0.0
  description: |
    REST API for Savezy, providing authentication workflows and expense management
    for authenticated users. All expense endpoints require a JWT Bearer token
    issued by the authentication flow.
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://api.savezy.com
    description: Production server (example)
...
paths:
...
  /api/expenses:
    get:
      tags: [Expenses]
      summary: List expenses
      description: Returns paginated expenses for the authenticated user with optional filters.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/CategoryFilter'
        - $ref: '#/components/parameters/SortField'
        - $ref: '#/components/parameters/SortOrder'
      responses:
        '200':
          description: Expenses retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    post:
      tags: [Expenses]
      summary: Create an expense
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseRequest'
      responses:
        '201':
          description: Expense created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseSingleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/expenses/{expense_id}:
    parameters:
      - name: expense_id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Expenses]
      summary: Retrieve a single expense
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Expense retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseSingleResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Expenses]
      summary: Update an expense
      description: Applies partial updates to an expense owned by the user.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateExpenseRequest'
      responses:
        '200':
          description: Expense updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseSingleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Expenses]
      summary: Delete an expense
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Expense deleted successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Expense deleted successfully.
                  data:
                    type: object
                    properties:
                      expense_id:
                        type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      description: Page number (1-indexed).
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      description: Number of items per page (1-100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    CategoryFilter:
      name: category
      in: query
      description: Filter expenses by category slug.
      schema:
        type: string
        enum: [investment, wants, need]
    SortField:
      name: sort
      in: query
      description: Field used for sorting.
      schema:
        type: string
        enum: [date, amount, title, category, card]
        default: date
    SortOrder:
      name: order
      in: query
      description: Sort order.
      schema:
        type: string
        enum: [asc, desc]
        default: desc
  responses:
    BadRequest:
      description: Invalid request or validation error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: Rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        title:
          type: string
        amount:
          type: number
          format: float
        category:
          type: string
          description: Category slug (investment, wants, need).
        category_name:
          type: string
          description: Human readable category label.
        date:
          type: string
          format: date-time
        description:
          type: string
          nullable: true
        card:
          allOf:
            - $ref: '#/components/schemas/Card'
          nullable: true
    Card:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
        brand:
          type: string
          nullable: true
        last_four:
          type: string
          nullable: true
...
    ExpenseListResponse:
      type: object
      properties:
        message:
          type: string
          example: Expenses retrieved successfully.
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Expense'
            pagination:
              type: object
              properties:
                page:
                  type: integer
                limit:
                  type: integer
                total_pages:
                  type: integer
                total_items:
                  type: integer
                has_next:
                  type: boolean
                has_prev:
                  type: boolean
            filters:
              type: object
              properties:
                category:
                  type: string
                  nullable: true
                sort:
                  type: string
                order:
                  type: string
    ExpenseSingleResponse:
      type: object
      properties:
        message:
          type: string
          example: Expense created successfully.
        data:
          type: object
          properties:
            expense:
              $ref: '#/components/schemas/Expense'
    CreateExpenseRequest:
      type: object
      required:
        - title
        - amount
        - category
        - card_id
      properties:
        title:
          type: string
        amount:
          type: number
          format: float
        category:
          type: string
          enum: [investment, wants, need]
        description:
          type: string
          nullable: true
        card_id:
          type: integer
        date:
          type: string
          format: date-time
          nullable: true
    UpdateExpenseRequest:
      type: object
      properties:
        title:
          type: string
        amount:
          type: number
          format: float
        category:
          type: string
          enum: [investment, wants, need]
        description:
          type: string
          nullable: true
        card_id:
          type: integer
        date:
          type: string
          format: date-time
          nullable: true
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Validation failed.
        data:
          type: object
          additionalProperties: true